Added support for MQTT retain and rewritten in-memory storage (#190)
This PR adds support for MQTT Retain flag. In order to make retained messages play well with emitter's message history, we consider that the retained messages are just messages with a large TTL (defaults to `30 days` ). The `retain` TTL is configurable in the storage provider, so you can setup a longer or shorter time.

1. The retain flag will only work if the key has `store` permission
2. The subscriptions with `load` permission will default to `last=1` . This is done in order to be compliant with the MQTT spec as the retain message should always be sent to the subscriptions.

Now, you can still have `ttl` and `last` options and it should play well with the retained messages, sending everything.