Allow extending keys through keygen (#268)
This adds a way to extend keys through a normal `keygen` method. Users will need to provide a key which has `extend` permissions in the request along with the desired access permissions. Note that the permissions will be a reflection of the key itself (e.g: if a key has `re` permissions and `keygen` request contains `wr`, the resulting key will only have `r` permission).